[#25] refactor: AuthProvider, AuthGuard 역할 분리
