+++ src/context/AuthProvider.tsx
@@ -1,36 +1,29 @@
-import { createContext, useEffect, useState } from 'react'
+import { Dispatch, SetStateAction, createContext, useState } from 'react'
 
-import { User, onAuthStateChanged } from 'firebase/auth'
-
-import { auth } from '@/remote/firebase'
+import { User } from 'firebase/auth'
 
 interface AuthContextType {
   user: User | null
   initialized: boolean
+  setUser: Dispatch<SetStateAction<User | null>>
+  setInitialized: Dispatch<SetStateAction<boolean>>
 }
 
 export const AuthContext = createContext<AuthContextType>({
   user: null,
   initialized: false,
+  setUser: () => {},
+  setInitialized: () => {},
 })
 
 function AuthProvider({ children }: { children: React.ReactNode }) {
   const [user, setUser] = useState<User | null>(null)
   const [initialized, setInitialized] = useState(false)
 
-  useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
-      setUser(currentUser)
-      setInitialized(true)
-    })
-
-    return () => {
-      unsubscribe()
-    }
-  }, [])
-
   return (
-    <AuthContext.Provider value={{ user, initialized }}>
+    <AuthContext.Provider
+      value={{ user, initialized, setUser, setInitialized }}
+    >
       {children}
     </AuthContext.Provider>
   )
