+++ src/hooks/useAuth.tsx
@@ -3,6 +3,5 @@ import { useContext } from 'react'
 import { AuthContext } from '@/context/AuthProvider'
 
 export const useAuth = () => {
-  const auth = useContext(AuthContext)
-  return auth
+  return useContext(AuthContext)
 }
